LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_origin-expected.txt

Blocked access to external URL http://www.localhost:8800/resource-timing/resources/TAOResponse.py?tao=origin
CONSOLE MESSAGE: line 13: NetworkError:  A network error occurred.
Description

This test validates that for a cross origin resource, the timing allow check algorithm will pass when the value of Timing-Allow-Origin is a case-sensitive match for the value of the origin of the current document.


PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
